IPA rpath
When we build a shared library for testing, it is built differently
whether it is meant for the local system or a remote one.  When it is
for the local system, the library is built with no SONAME.  So when the
executable is built, roughly in this way:

  $ gcc testfile.c /path/to/library.so

the executable will contain an absolute reference to the library.  For
example:

  $ readelf -a testsuite/gdb.python/py-shared | grep NEEDED
   0x0000000000000001 (NEEDED)             Shared library: [/home/emaisin/build/binutils-gdb/gdb/testsuite/gdb.python/py-shared-sl.sl]

When testing is done remotely, the absolute path obviously doesn't work.
Therefore, we build the library with an SONAME:

  $ readelf -a testsuite/gdb.python/py-shared-sl.sl | grep SONAME
   0x000000000000000e (SONAME)             Library soname: [py-shared-sl.sl]

which ends up in the executable's NEEDED field:

  $ readelf -a testsuite/gdb.python/py-shared | grep NEEDED
   0x0000000000000001 (NEEDED)             Shared library: [py-shared-sl.sl]

To allow the dynamic linker to find the shared library on the target, we
have to add the special RPATH value $ORIGIN, which tells it to search in
the executable's directory:

  $ readelf -a testsuite/gdb.python/py-shared | grep ORIGIN
   0x000000000000000f (RPATH)              Library rpath: [$ORIGIN]

TODO

problem with libipa is that it doesn't have an SONAME.  So when we link
an executable to it using full name, we can't run it on a remote
system.  I tried to change the tests to link with -L${ipa_dir}
-linproctrace, but it's cumbersome.  And we still need to add rpath
$ORIGIN.

We don't have control on IPA library in the testsuite.  For it to work
on remote systems, we need it to have an SONAME, we could add it.  But
now it doesn't work with native/local systems, because the executable
doesn't have the full path to the library anymore.  So I suggest always
building shlibs with SONAME, and always adding RPATH ORIGIN to test
executables.  Modify gdb_load_shlibs to just copy the files when native,
if dest != source.