Fix structure alignment problems with IPA protocol objects.
While testing fast tracepoints on ARM I came across this problem :

Program received signal SIGSEGV, Segmentation fault.
0x4010b56c in ?? () from target:/lib/arm-linux-gnueabihf/libc.so.6
(gdb) FAIL: gdb.trace/ftrace.exp: ond globvar > 7: continue

The problem is that on GDBServer side struct tracepoint_action is aligned
on 4 bytes, and collect_memory_action is aligned on 8. However in the IPA
they are both aligned on 8 bytes.

Thus when we copy the data from GDBServer's struct tracepoint_action->type
offset to the ipa the alignement is wrong and the addr,len,basereg values
are wrong also, causing a crash in the inferiror as it tries to read
memory at a bogus address.

This patch fixes this issue by setting the tracepoint_action and
collect_memory_action as packed.

This patch also fixes this issue in general by setting all IPA protocol
object structures as packed.